retire the old matrix.h and start using d3math. Remove all use of Vector3d, QuaternionD... use the Vector3f, QuaternionF, ... instead.

+++ source/cortex-m4/sdfirmware/src/pidtorque.cpp
@@ -68,7 +68,7 @@ Vector3f PidTorque::get_torque(const QuaternionF &in_Q, const TimeSpan& dt, floa
 	// ==> errQ = (~attitudeQ) * targetQ;
 	Qtorq = (~in_Q) * set_Q_;
 	QuaternionF swing;
-	QuaternionF::decomposeTwistSwing(Qtorq, Vector3d(0,0,1), swing, twist_);
+	QuaternionF::decomposeTwistSwing(Qtorq, Vector3f(0,0,1), swing, twist_);
 	Vector3f error_z = twist_.axis().normalize() * twist_.angle() * 1.0;
 	error_z.at(0) = error_z.at(1) = 0.0;
 	torq.at(2) = pid_controller_yaw_.get_d(error_z, dt.seconds_float()).at(2);
